Attended PR comments
