+++ framework-hooks/framework-photoshop-bootstrap/source/ftrack_framework_photoshop_bootstrap/__init__.py
@@ -1,10 +1,19 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014-2023 ftrack
 
-from ftrack_framework_photoshop.bootstrap import initialise, run
+from ftrack_framework_photoshop import bootstrap_integration, run_integration
 
 # Initialise the Photoshop Framework Python standalone part.
-initialise()
+bootstrap_integration(
+    panel_launchers=[
+        {
+            "name": "publish",
+            "label": "PUBLISHER",
+            "dialog_name": "framework_publisher_dialog",
+            "image": "publish",
+        }
+    ]
+)
 
 # Run until Photoshop is closed.
-run()
+run_integration()
